Added polynomial extrapolation and gradient calculation (#5)
* Added polynomial extrapolation and gradient calculation

Since I need simple polynomial extrapolation for splines, I added a corresponding subroutine. I also added the possibility to calculate gradients and replaced the while-loop used to find the correct interval for a given value with a binary search which should be faster on average. Those features are documented in the Readme and I also extended the test suite accordingly.